Split representations.rs into Expressions module with multiple files
