+++ src/DataReader/serialization.rs
@@ -2,7 +2,7 @@ use crate::DataReader::parse_edge;
 use crate::ModelObjects::component::{
     Component, Declarations, Edge, Location, LocationType, SyncType,
 };
-use crate::ModelObjects::representations;
+use crate::ModelObjects::Expressions;
 use crate::Simulation::graph_layout::layout_dummy_component;
 use edbm::util::constraints::ClockIndex;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
@@ -47,7 +47,7 @@ pub struct DummyEdge {
         deserialize_with = "decode_guard",
         serialize_with = "encode_opt_boolexpr"
     )]
-    pub guard: Option<representations::BoolExpression>,
+    pub guard: Option<Expressions::BoolExpression>,
     #[serde(
         deserialize_with = "decode_update",
         serialize_with = "encode_opt_updates"
@@ -140,7 +140,7 @@ pub struct DummyLocation {
         //deserialize_with = "decode_invariant",
         serialize_with = "encode_opt_boolexpr"
     )]
-    pub invariant: Option<representations::BoolExpression>,
+    pub invariant: Option<Expressions::BoolExpression>,
     #[serde(
         //deserialize_with = "decode_location_type",
         serialize_with = "encode_location_type",
@@ -232,7 +232,7 @@ where
 /// Function used for deserializing guards
 pub fn decode_guard<'de, D>(
     deserializer: D,
-) -> Result<Option<representations::BoolExpression>, D::Error>
+) -> Result<Option<Expressions::BoolExpression>, D::Error>
 where
     D: Deserializer<'de>,
 {
@@ -263,7 +263,7 @@ where
 //Function used for deserializing invariants
 pub fn decode_invariant<'de, D>(
     deserializer: D,
-) -> Result<Option<representations::BoolExpression>, D::Error>
+) -> Result<Option<Expressions::BoolExpression>, D::Error>
 where
     D: Deserializer<'de>,
 {
@@ -368,7 +368,7 @@ where
 }
 
 pub fn encode_opt_boolexpr<S>(
-    opt_expr: &Option<representations::BoolExpression>,
+    opt_expr: &Option<Expressions::BoolExpression>,
     serializer: S,
 ) -> Result<S::Ok, S::Error>
 where
@@ -382,7 +382,7 @@ where
 }
 
 pub fn encode_boolexpr<S>(
-    expr: &representations::BoolExpression,
+    expr: &Expressions::BoolExpression,
     serializer: S,
 ) -> Result<S::Ok, S::Error>
 where
@@ -392,7 +392,7 @@ where
 }
 
 pub fn encode_arithexpr<S>(
-    expr: &representations::ArithExpression,
+    expr: &Expressions::ArithExpression,
     serializer: S,
 ) -> Result<S::Ok, S::Error>
 where
